# Use a Javascript Map for Go Maps.  Improves performance of len() calls by orders of magnitude. 🗺️ 