--- a/compiler/expressions.go
+++ b/compiler/expressions.go
@@ -475,9 +475,21 @@ func (fc *funcContext) translateExpr(expr ast.Expr) *expression {
 			}
 			key := fmt.Sprintf("%s.keyFor(%s)", fc.typeName(t.Key()), fc.translateImplicitConversion(e.Index, t.Key()))
 			if _, isTuple := exprType.(*types.Tuple); isTuple {
-				return fc.formatExpr(`(%1s = %2e[%3s], %1s !== undefined ? [%1s.v, true] : [%4e, false])`, fc.newVariable("_entry"), e.X, key, fc.zeroValue(t.Elem()))
+				return fc.formatExpr(
+					`(%1s = $mapIndex(%2e,%3s), %1s !== undefined ? [%1s.v, true] : [%4e, false])`,
+					fc.newVariable("_entry"),
+					e.X,
+					key,
+					fc.zeroValue(t.Elem()),
+				)
 			}
-			return fc.formatExpr(`(%1s = %2e[%3s], %1s !== undefined ? %1s.v : %4e)`, fc.newVariable("_entry"), e.X, key, fc.zeroValue(t.Elem()))
+			return fc.formatExpr(
+				`(%1s = $mapIndex(%2e,%3s), %1s !== undefined ? %1s.v : %4e)`,
+				fc.newVariable("_entry"),
+				e.X,
+				key,
+				fc.zeroValue(t.Elem()),
+			)
 		case *types.Basic:
 			return fc.formatExpr("%e.charCodeAt(%f)", e.X, e.Index)
 		default:
@@ -919,9 +931,9 @@ func (fc *funcContext) translateBuiltin(name string, sig *types.Signature, args 
 			return fc.formatExpr("$makeSlice(%s, %f)", t, args[1])
 		case *types.Map:
 			if len(args) == 2 && fc.pkgCtx.Types[args[1]].Value == nil {
-				return fc.formatExpr(`((%1f < 0 || %1f > 2147483647) ? $throwRuntimeError("makemap: size out of range") : {})`, args[1])
+				return fc.formatExpr(`((%1f < 0 || %1f > 2147483647) ? $throwRuntimeError("makemap: size out of range") : new $global.Map())`, args[1])
 			}
-			return fc.formatExpr("{}")
+			return fc.formatExpr("new $global.Map()")
 		case *types.Chan:
 			length := "0"
 			if len(args) == 2 {
@@ -940,7 +952,7 @@ func (fc *funcContext) translateBuiltin(name string, sig *types.Signature, args 
 		case *types.Pointer:
 			return fc.formatExpr("(%e, %d)", args[0], argType.Elem().(*types.Array).Len())
 		case *types.Map:
-			return fc.formatExpr("$keys(%e).length", args[0])
+			return fc.formatExpr("(%e ? %e.size : 0)", args[0], args[0])
 		case *types.Chan:
 			return fc.formatExpr("%e.$buffer.length", args[0])
 		// length of array is constant
@@ -969,7 +981,12 @@ func (fc *funcContext) translateBuiltin(name string, sig *types.Signature, args 
 	case "delete":
 		args = fc.expandTupleArgs(args)
 		keyType := fc.pkgCtx.TypeOf(args[0]).Underlying().(*types.Map).Key()
-		return fc.formatExpr(`delete %e[%s.keyFor(%s)]`, args[0], fc.typeName(keyType), fc.translateImplicitConversion(args[1], keyType))
+		return fc.formatExpr(
+			`$mapDelete(%1e, %2s.keyFor(%3s))`,
+			args[0],
+			fc.typeName(keyType),
+			fc.translateImplicitConversion(args[1], keyType),
+		)
 	case "copy":
 		args = fc.expandTupleArgs(args)
 		if basic, isBasic := fc.pkgCtx.TypeOf(args[1]).Underlying().(*types.Basic); isBasic && isString(basic) {