--- a/compiler/natives/doc.go
+++ b/compiler/natives/doc.go
@@ -5,5 +5,9 @@
 // in src subfolder.
 package natives
 
+import (
+	_ "github.com/shurcooL/vfsgen" // Force go.mod to require this package
+)
+
 //go:generate vfsgendev -source="github.com/gopherjs/gopherjs/compiler/natives".FS -tag=gopherjsdev
 //go:generate gofmt -w -s fs_vfsdata.go