… compiler/gopherjspkg/fs_vfsdata.go
… compiler/natives/fs_vfsdata.go
… compiler/prelude/prelude_min.go