# Overview
https://github.com/gopherjs/gopherjs/issues/1135

The performance of len() on maps brought me here, because it would call js `Object.keys(myMap).Length`.  It was many orders of magnitude slower than len() in Go.  It was a pitfall that those writing idiomatic go would fall into.

This PR switches the backing implementation of Maps in GopherJS to [ECMAScript 2015 Maps](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Map).  These maps provide accounting so that Map.size can be called to make len() fast.  I was hopeful that iterating maps would be faster, also, because the len of the map is used, but unfortunately that is not the case.  However, it should be trivial to change the loop implementation to use `[Map.entries()`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Map/entries) when GopherJS adopts ECMAScript 2017.  


One thing to note:  ES 2015 Maps do not return in random order.  There is a pitfall, the I believe is acceptable, that someone could write GopherJS code that relies on the order of elements in a map, that does not work when run in the Go runtime.

Using the benchmarks committed as part of this PR, with a size of 10000 elements:
**Master**
```
✗ GOOS=linux gopherjs test ./... --bench="Benchmark.*"
Using GOOS=linux and GOARCH=ecmascript in GopherJS is deprecated and will be removed in future. Use GOOS=js GOARCH=ecmascript instead.
goos: js
goarch: ecmascript
BenchmarkSliceLen           	1000000000	         0.2600 ns/op
BenchmarkMapLen             	    9886	    115719 ns/op
BenchmarkMapNilCheck        	1000000000	         0.5070 ns/op
BenchmarkMapNilElementCheck 	144578312	         8.328 ns/op
BenchmarkSliceRange         	  226414	      5070 ns/op
BenchmarkMapRange           	    6315	    183215 ns/op
PASS
ok  	benchmark/tom	7.682s
```

**This Branch**
Map len is ~6 orders of magnitude faster.  Unfortunately, it appears element access and ranging has gotten ~20% worse, and ranging ~3% worse
```
✗ GOOS=linux gopherjs test ./... --bench="Benchmark.*"
Using GOOS=linux and GOARCH=ecmascript in GopherJS is deprecated and will be removed in future. Use GOOS=js GOARCH=ecmascript instead.
goos: js
goarch: ecmascript
BenchmarkSliceLen           	1000000000	         0.2820 ns/op
BenchmarkMapLen             	1000000000	         0.5100 ns/op
BenchmarkMapNilCheck        	1000000000	         0.2610 ns/op
BenchmarkMapNilElementCheck 	104008665	        10.50 ns/op
BenchmarkSliceRange         	  222222	      5094 ns/op
BenchmarkMapRange           	    6082	    188096 ns/op
PASS
ok benchmark/tom	7.001s
```

# Testing
It seems like the tests cover many more cases than manual testing ever could.  I may have tunnel vision as the author of the PR, but I believe that passing CI tests should be considered good enough.