@tomconnell-wf 2022-08-05
At this point, maps cannot be accessed by index, but their size can be acquired, and they can be iterated.

Using the same benchmark I used to report the slowness (with a million entries):
```
goos: linux
goarch: js
BenchmarkSliceLen           	1000000000	         0.2530 ns/op
BenchmarkMapLen             	1000000000	         0.5040 ns/op
BenchmarkMapNilCheck        	1000000000	         0.5050 ns/op
BenchmarkMapNilElementCheck 	144578312	         8.300 ns/op
PASS
```

The len check is as quick as I would hope.  (The setup of the map is very slow, and could possibly be improved (Because I'm making a slice of maps that then get converted in $makeMap, but I don't think that is any different than master.)