@tomconnell-wf 2022-08-05
FWIW, this is with 1 million elements
```
goos: linux
goarch: js
BenchmarkSliceLen           	1000000000	         0.2560 ns/op
BenchmarkMapLen             	1000000000	         0.5920 ns/op
BenchmarkMapNilCheck        	1000000000	         0.6890 ns/op
BenchmarkMapNilElementCheck 	98797087	        10.13 ns/op
BenchmarkSliceRange         	    2221	    510581 ns/op
```