@tomconnell-wf 2022-08-08
Well, I think I am almost there.  It looks like I have a Sprintf problem to fix, and then done?