@tomconnell-wf 2022-08-09
As you can see in CI, there are some fmt stdlib tests that are failing.  It seems like it is because I didn't wire something up to reflect/reflectlite properly.  

Here's one of the tests that are failing.
https://cs.opensource.google/go/go/+/master:src/fmt/example_test.go;l=295-300?q=dach&ss=go%2Fgo

The expected output is
```
map[dachshund:false peanut:true] map[string]bool{"dachshund":false, "peanut":true}
```

The incorrect output is 
```
map[] map[string]bool{}
```

I'm a bit out of my depth here, but it seems like this is because the keys and values aren't reflecting properly.  My gut says it has something to do with the change from e to e.v in https://github.com/gopherjs/gopherjs/pull/1136/files#diff-70e5ca5fb5a75cb3a234c5a2caee3036426e4ffbb46a9e6b2ee556cb244eccbcR623.  I'll keep taking cracks at it, but if the maintainers have a hunch, I could use the guidance.