@tomconnell-wf 2022-08-09
I'm going with the theory that by doing [this](https://github.com/gopherjs/gopherjs/pull/1136/files#diff-70e5ca5fb5a75cb3a234c5a2caee3036426e4ffbb46a9e6b2ee556cb244eccbcR623) I clobbered the information needed for reflection to work properly.  I'll massage it so that we can use the Map as a POJSO, with the reflection type info...I think.