@nevkontakte 2022-08-09
@tomconnell-wf I've looked through the diff and there is nothing obviously incorrect in it I could pin the test failure on. My best guess is that somewhere in the `reflect` package we read the `$keys().length` value, which is now 0, and so iteration is not happening properly. I am not sure I will have time for a deep dive until the weekend, but if you don't find the bug until then, I'll give it a shot.

As a side note, I recommend you rebasing on top of the latest master, which includes Go 1.18 support. There have been some additions to the `reflect` map-related APIs in 1.18 which may be impacted by your change or will cause merge conflicts when you are ready.