@flimzy 2022-08-10
> I think the official recommendation it to create a source tools.go somewhere

We can go one step further, and put it in a subdir so that its in its own package that is never even parsed during compilation (but is parsed by `go mod`)