@tomconnell-wf 2022-08-10
This may be considered phoning it in, but I can't figure out the reflection.  As best I can understand, it's built around the keys and the elements being available on the javascript object.  Rather than having a special case for Map types, and since these are all references anyway, I am thinking of leaving the key and element as they were before, but also putting them into the javascript Map api, for faster size calls.  I will push commits for this later today, and some javascript heap snapshots to make sure memory usage isn't significantly higher.