@nevkontakte 2022-08-10
I would strongly prefer not to do this, not for performance, but for maintenance reasons. A lot of our work on gopherjs over the past year has been dedicated to simplifying and documenting the codebase, and I would very much prefer to help you figure out the reflection side of the problem over adding more cruft that will be confusing people in future.