@tomconnell-wf 2022-08-10
Side note - when it is decided that ECMA Script 2017 is supported, we can use `.entries()` instead of `.keys()` on the map range statement, and it'll be about 2x faster.  
https://github.com/gopherjs/gopherjs/commit/cf9a8fa87bb80f90e95df2c21e13a50e85d96d49

10000 elements:

**Before**
✗ GOOS=linux gopherjs test ./... --bench="Benchmark.*"
Using GOOS=linux and GOARCH=ecmascript in GopherJS is deprecated and will be removed in future. Use GOOS=js GOARCH=ecmascript instead.
goos: js
goarch: ecmascript
BenchmarkSliceLen           	1000000000	         0.2520 ns/op
BenchmarkMapLen             	1000000000	         0.2490 ns/op
BenchmarkMapNilCheck        	1000000000	         0.5010 ns/op
BenchmarkMapNilElementCheck 	363636363	         3.311 ns/op
BenchmarkSliceRange         	  222222	      4991 ns/op
BenchmarkMapRange           	    5454	    186285 ns/op
PASS

**After**
✗ GOOS=linux gopherjs test ./... --bench="Benchmark.*"
Using GOOS=linux and GOARCH=ecmascript in GopherJS is deprecated and will be removed in future. Use GOOS=js GOARCH=ecmascript instead.
goos: js
goarch: ecmascript
BenchmarkSliceLen           	1000000000	         0.2440 ns/op
BenchmarkMapLen             	1000000000	         0.2410 ns/op
BenchmarkMapNilCheck        	1000000000	         0.4830 ns/op
BenchmarkMapNilElementCheck 	377358489	         3.246 ns/op
BenchmarkSliceRange         	  222222	      4847 ns/op
BenchmarkMapRange           	   13656	     83553 ns/op
PASS