@tomconnell-wf 2022-08-10
> I would strongly prefer not to do this, not for performance, but for maintenance reasons
I spent most of the day yesterday trying to figure it out.  I'll give it another shot.  Maybe having slept  on it will unlock something.