@tomconnell-wf 2022-08-10
Current state of performance.  Range is a bit worse?!?  :(

10000 elements

Master:
```
✗ GOOS=linux gopherjs test ./... --bench="Benchmark.*"
Using GOOS=linux and GOARCH=ecmascript in GopherJS is deprecated and will be removed in future. Use GOOS=js GOARCH=ecmascript instead.
goos: js
goarch: ecmascript
BenchmarkSliceLen           	1000000000	         0.2500 ns/op
BenchmarkMapLen             	   10774	    107203 ns/op
BenchmarkMapNilCheck        	1000000000	         0.4900 ns/op
BenchmarkMapNilElementCheck 	150753768	         8.026 ns/op
BenchmarkSliceRange         	  230768	      4853 ns/op
BenchmarkMapRange           	    6315	    164054 ns/op
PASS
ok  	github.com/Workiva/doc_plat_client/tool/internal/gopherjsbenchmarks/benchmark/tom	7.427s
```


This commit:
```
✗ GOOS=linux gopherjs test ./... --bench="Benchmark.*"
Using GOOS=linux and GOARCH=ecmascript in GopherJS is deprecated and will be removed in future. Use GOOS=js GOARCH=ecmascript instead.
goos: js
goarch: ecmascript
BenchmarkSliceLen           	1000000000	         0.2440 ns/op
BenchmarkMapLen             	1000000000	         0.2430 ns/op
BenchmarkMapNilCheck        	1000000000	         0.4840 ns/op
BenchmarkMapNilElementCheck 	357142856	         3.214 ns/op
BenchmarkSliceRange         	  218181	      4968 ns/op
BenchmarkMapRange           	    6458	    183958 ns/op
PASS
ok  	github.com/Workiva/doc_plat_client/tool/internal/gopherjsbenchmarks/benchmark/tom	6.549s
```