@tomconnell-wf 2022-08-10
Those numbers make me think I could change tack and just account for the size of the map during creating, assign and delete.  That seems semantically less correct, through. 🤔